backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
faiss-cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.28.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1

python-dotenv
sentence-transformers
faiss-cpu
google-genai